requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
python-multipart==0.0.9
jinja2==3.1.4
anthropic==0.84.0
PyMuPDF==1.26.5
python-docx==1.2.0
python-pptx==1.0.2
openpyxl==3.1.5
aiofiles==23.2.1
